Refactor SaveCrisisTreatmentUseCase.
