+++ app/src/main/java/com/devmob/alaya/domain/CrisisTreatmentRepository.kt
@@ -4,6 +4,6 @@ import com.devmob.alaya.domain.model.FirebaseResult
 import com.devmob.alaya.domain.model.OptionTreatment
 
 interface CrisisTreatmentRepository {
-    suspend fun addCustomTreatment(patientEmail: String, treatment: List<OptionTreatment?>) : FirebaseResult
+    suspend fun saveCustomTreatment(patientEmail: String, treatment: List<OptionTreatment?>) : FirebaseResult
     suspend fun getCustomTreatment(patientEmail: String) : List<OptionTreatment>?
 }
